+++ R/plot_volcano.R
@@ -47,7 +47,6 @@
 #'
 #' se <- airway
 #' se$offset <- log(colSums(SummarizedExperiment::assay(se, "counts")))
-#' se <- tidybulk::estimate_dispersion(se, formula_abundance = ~ dex + cell)
 #' se <- estimate_dispersion_prior(se, formula_abundance = ~ dex + cell)
 #'
 #' pipeline <- se |>